refactor(specs): added separate spec files for the core concepts of the plugin
